requirements.txt
alembic==1.4.3
astroid==2.4.2
attrs==20.2.0
Authlib==0.14.3
bandit==1.6.2
boto3==1.15.3
botocore==1.18.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coverage==5.3
cryptography==3.1.1
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-CDN==1.5.3
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
Flask-WTF==0.14.3
gitdb==4.0.5
GitPython==3.1.8
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pbr==5.5.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.19
stevedore==3.2.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.2.0
